backend/requirements.txt
asgiref==3.8.1
setuptools==68.0.0
Django==5.2.2
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
environ==1.0
pillow==11.2.1
sqlparse==0.5.3
gunicorn==21.2.0
